webshoplanding (no header/nav or footer)
We need to put it all together as a coming step. 

US: as a developer I want the created header, content, footer to be merged into html-files and css-files so we can work from them as templates. 

Would also like my work reviewed so we can re-do if it doesn't hold up.